crawler_poc/requirements.txt
redis
requests
psycopg2-binary
minio
playwright
python-dotenv
beautifulsoup4
lxml
